requirements.txt
django
djangorestframework
psycopg2-binary
web3
python-decouple
cryptography
